// src/domain/checkout/types/country.ts
export type Region =
  | "Slovenia"
  | "EU"
  | "Non-EU Europe"
  | "UK"
  | "North America"
  | "Australia"
  | "Japan";

export type CountryCode =
  // EU Countries
  | "AT"
  | "BE"
  | "BG"
  | "HR"
  | "CY"
  | "CZ"
  | "DK"
  | "EE"
  | "FI"
  | "FR"
  | "DE"
  | "GR"
  | "HU"
  | "IE"
  | "IT"
  | "LV"
  | "LT"
  | "LU"
  | "MT"
  | "NL"
  | "PL"
  | "PT"
  | "RO"
  | "SK"
  | "SI"
  | "ES"
  | "SE"

  // Non-EU Europe
  | "UK"
  | "RS"
  | "BA"
  | "ME"
  | "MK"

  // Other regions
  | "US"
  | "CA"
  | "AU"
  | "JP";

export type Country = {
  code: CountryCode;
  region: Region;
  name: string;
  flag: string;
  callingCode: string;
  orderPriority?: number;
};